chore: refactor: internal smooth transitions.
* Moved the transitionHandler into its own file.
* Use the new customOptions property in ccg-state to pass through parameters
